+++ Assets/RemoveEdge.ps1
@@ -1,288 +1,669 @@
-$also_remove_webview = 0
-## why also remove webview? because it is 2 copies of edge, not a slimmed down CEF, and is driving bloated web apps
-$also_remove_widgets = 0
-## why also remove widgets? because it is a webview glorified ad portal on msn and bing news cathering to stupid people
-$also_remove_xsocial = 0
-## why also remove xsocial? because it starts webview setup every boot - xbox gamebar will still work without the social crap
+#Requires -Version 5.0
 
-$host.ui.RawUI.WindowTitle = 'Edge Removal - RyTuneX'
-$remove_appx = @("MicrosoftEdge"); $remove_win32 = @("Microsoft Edge","Microsoft Edge Update"); $skip = @("DevTools", "MicrosoftEdge")
-if ($also_remove_webview -eq 1) {$remove_appx += "Win32WebViewHost"; $remove_win32 += "Microsoft EdgeWebView"}
-if ($also_remove_widgets -eq 1) {$remove_appx += "WebExperience"}
-if ($also_remove_xsocial -eq 1) {$remove_appx += "GamingServices"}
+<#
+	.SYNOPSIS
+	Uninstalls or reinstalls Microsoft Edge and its related components. Made by @he3als.
 
-$global:WEBV = $also_remove_webview -eq 1
-$global:IS64 = [Environment]::Is64BitOperatingSystem
-$global:IFEO = 'HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Image File Execution Options'
-$global:EDGE_UID = '{56EB18F8-B008-4CBD-B6D2-8C97FE7E9062}'
-$global:WEBV_UID = '{F3017226-FE2A-4295-8BDF-00C3A9A7E4C5}'
-$global:UPDT_UID = '{F3C4FE00-EFD5-403B-9569-398A20F1BA4A}'
-$global:PROGRAMS = ($env:ProgramFiles, ${env:ProgramFiles(x86)})[$IS64]
-$global:SOFTWARE = ('SOFTWARE', 'SOFTWARE\WOW6432Node')[$IS64]
-$global:ALLHIVES = 'HKCU:\SOFTWARE','HKLM:\SOFTWARE','HKCU:\SOFTWARE\Policies','HKLM:\SOFTWARE\Policies'
-if ($IS64) { $global:ALLHIVES += "HKCU:\$SOFTWARE","HKLM:\$SOFTWARE","HKCU:\$SOFTWARE\Policies","HKLM:\$SOFTWARE\Policies"}
-## -------------------------------------------------------------------------------------------------------------------------------
+	.Description
+	Uninstalls or reinstalls Microsoft Edge and its related components in a non-forceful manner, based upon switches or user choices in a TUI.
 
-## 1 bonus! enter into powershell console: firefox / edge / webview to install a browser / reinstall edge / webview after removal
-function global:firefox { $url = 'https://download.mozilla.org/?product=firefox-stub'
-  $setup = "$((new-object -ComObject Shell.Application).NameSpace('shell:Downloads').Self.Path)\Firefox Installer.exe"
-  write-host $url; Invoke-WebRequest $url -OutFile $setup; start $setup
+	.PARAMETER UninstallEdge
+	Uninstalls Edge, leaving the Edge user data.
+
+	.PARAMETER InstallEdge
+	Installs Edge, leaving the previous Edge user data.
+
+	.PARAMETER InstallWebView
+	Installs Edge WebView2 using the Evergreen installer.
+
+	.PARAMETER RemoveEdgeData
+	Removes all Edge user data. Compatible with -InstallEdge.
+
+	.PARAMETER KeepAppX
+	Doesn't check for and remove the AppX, in case you want to use alternative AppX removal methods. Doesn't work with UninstallEdge.
+
+	.PARAMETER NonInteractive
+	When combined with other parameters, this does not prompt the user for anything.
+
+	.LINK
+	https://github.com/he3als/EdgeRemover
+#>
+
+param (
+	[switch]$UninstallEdge,
+	[switch]$InstallEdge,
+	[switch]$InstallWebView,
+	[switch]$RemoveEdgeData,
+	[switch]$KeepAppX,
+	[switch]$NonInteractive
+)
+
+$version = '1.9.4'
+
+$ProgressPreference = "SilentlyContinue"
+$sys32 = [Environment]::GetFolderPath('System')
+$windir = [Environment]::GetFolderPath('Windows')
+$env:path = "$windir;$sys32;$sys32\Wbem;$sys32\WindowsPowerShell\v1.0;" + $env:path
+$baseKey = "HKLM:\SOFTWARE" + $(if ([Environment]::Is64BitOperatingSystem) { "\WOW6432Node" }) + "\Microsoft"
+$msedgeExe = "$([Environment]::GetFolderPath('ProgramFilesx86'))\Microsoft\Edge\Application\msedge.exe"
+$edgeUWP = "$windir\SystemApps\Microsoft.MicrosoftEdge_8wekyb3d8bbwe"
+
+if ($NonInteractive -and (!$UninstallEdge -and !$InstallEdge -and !$InstallWebView)) {
+	$NonInteractive = $false
 }
-function global:edge { $url = 'https://go.microsoft.com/fwlink/?linkid=2108834&Channel=Stable&language=en'
-  $setup = "$((new-object -ComObject Shell.Application).NameSpace('shell:Downloads').Self.Path)\MicrosoftEdgeSetup.exe"
-  write-host $url; Invoke-WebRequest $url -OutFile $setup; PREPARE_EDGE; start $setup
-}
-function global:webview { $url = 'https://go.microsoft.com/fwlink/p/?LinkId=2124703'
-  $setup = "$((new-object -ComObject Shell.Application).NameSpace('shell:Downloads').Self.Path)\MicrosoftEdgeWebview2Setup.exe"
-  write-host $url; Invoke-WebRequest $url -OutFile $setup; PREPARE_WEBVIEW; start $setup
-}
-function global:xsocial { $url = 'https://dlassets-ssl.xboxlive.com/public/content/XboxInstaller/XboxInstaller.exe'
-  $setup = "$((new-object -ComObject Shell.Application).NameSpace('shell:Downloads').Self.Path)\XboxInstaller.exe"
-  write-host $url; Invoke-WebRequest $url -OutFile $setup; PREPARE_WEBVIEW; start $setup
-}
-
-## helper for set-itemproperty remove-itemproperty new-item remove-item with auto test-path
-function global:sp_test_path { if (test-path $args[0]) {Microsoft.PowerShell.Management\Set-ItemProperty @args} else {
-  Microsoft.PowerShell.Management\New-Item $args[0] -force -ea 0 >''; Microsoft.PowerShell.Management\Set-ItemProperty @args} }
-function global:rp_test_path { if (test-path $args[0]) {Microsoft.PowerShell.Management\Remove-ItemProperty @args} }
-function global:ni_test_path { if (-not (test-path $args[0])) {Microsoft.PowerShell.Management\New-Item @args} }
-function global:ri_test_path { if (test-path $args[0]) {Microsoft.PowerShell.Management\Remove-Item @args} }
-foreach ($f in 'sp','rp','ni','ri') {set-alias -Name $f -Value "${f}_test_path" -Scope Local -Option AllScope -force -ea 0}
-
-## helper for edgeupdate reinstall
-function global:PREPARE_UPDT($cdp='msedgeupdate', $uid=$UPDT_UID) {
-  foreach ($f in 'sp','rp','ni','ri') {set-alias -Name $f -Value "${f}_test_path" -Scope Local -Option AllScope -force -ea 0}
-  foreach ($sw in $ALLHIVES) { 
-    rp "$sw\Microsoft\EdgeUpdate" 'DoNotUpdateToEdgeWithChromium' -force -ea 0
-    rp "$sw\Microsoft\EdgeUpdate" 'UpdaterExperimentationAndConfigurationServiceControl' -force -ea 0
-    rp "$sw\Microsoft\EdgeUpdate" "InstallDefault" -force -ea 0
-    rp "$sw\Microsoft\EdgeUpdate" "Install${uid}" -force -ea 0
-    rp "$sw\Microsoft\EdgeUpdate" "EdgePreview${uid}" -force -ea 0
-    rp "$sw\Microsoft\EdgeUpdate" "Update${uid}" -force -ea 0
-    rp "$sw\Microsoft\EdgeUpdate\ClientState\*" 'experiment_control_labels' -force -ea 0 
-    ri "$sw\Microsoft\EdgeUpdate\Clients\${uid}\Commands" -recurse -force -ea 0
-    rp "$sw\Microsoft\EdgeUpdateDev\CdpNames" "$cdp-*" -force -ea 0
-    sp "$sw\Microsoft\EdgeUpdateDev" 'CanContinueWithMissingUpdate' 1 -type Dword -force
-    sp "$sw\Microsoft\EdgeUpdateDev" 'AllowUninstall' 1 -type Dword -force
-  }
-}
-## helper for edge reinstall - remove bundled OpenWebSearch redirector and edgeupdate policies
-function global:PREPARE_EDGE {
-  foreach ($f in 'sp','rp','ni','ri') {set-alias -Name $f -Value "${f}_test_path" -Scope Local -Option AllScope -force -ea 0}
-  PREPARE_UPDT 'msedge' $EDGE_UID; PREPARE_UPDT 'msedgeupdate' $UPDT_UID 
-  $MSEDGE = "$PROGRAMS\Microsoft\Edge\Application\msedge.exe"
-  ri "$IFEO\msedge.exe" -recurse -force; ri "$IFEO\ie_to_edge_stub.exe" -recurse -force
-  ri 'Registry::HKEY_Users\S-1-5-21*\Software\Classes\microsoft-edge' -recurse -force
-  sp 'HKLM:\SOFTWARE\Classes\microsoft-edge\shell\open\command' '(Default)' "`"$MSEDGE`" --single-argument %%1" -force
-  ri 'Registry::HKEY_Users\S-1-5-21*\Software\Classes\MSEdgeHTM' -recurse -force
-  sp 'HKLM:\SOFTWARE\Classes\MSEdgeHTM\shell\open\command' '(Default)' "`"$MSEDGE`" --single-argument %%1" -force
-}
-## helper for webview reinstall - restore webexperience (widgets) if available
-function global:PREPARE_WEBVIEW {
-  PREPARE_UPDT 'msedgewebview' $WEBV_UID; PREPARE_UPDT 'msedgeupdate' $UPDT_UID
-  $cfg = @{Register=$true; ForceApplicationShutdown=$true; ForceUpdateFromAnyVersion=$true; DisableDevelopmentMode=$true} 
-  dir "$env:SystemRoot\SystemApps\Microsoft.Win32WebViewHost*\AppxManifest.xml" -rec -ea 0 | Add-AppxPackage @cfg
-  dir "$env:ProgramFiles\WindowsApps\MicrosoftWindows.Client.WebExperience*\AppxManifest.xml" -rec -ea 0 | Add-AppxPackage @cfg
-  kill -name explorer -ea 0; if ((get-process -name 'explorer' -ea 0) -eq $null) {start explorer}
-}
-## -------------------------------------------------------------------------------------------------------------------------------
-
-## 2 enable admin privileges
-$D1=[uri].module.gettype('System.Diagnostics.Process')."GetM`ethods"(42) |where {$_.Name -eq 'SetPrivilege'} #`:no-ev-warn
-'SeSecurityPrivilege','SeTakeOwnershipPrivilege','SeBackupPrivilege','SeRestorePrivilege'|foreach {$D1.Invoke($null, @("$_",2))}
-## -------------------------------------------------------------------------------------------------------------------------------
-
-## 3 shut down edge & webview clone stuff
-cd $env:systemdrive; taskkill /im explorer.exe /f 2>&1 >''
-$shut = 'explorer','Widgets','widgetservice','msedgewebview2','MicrosoftEdge*','chredge','msedge','edge'
-$shut,'msteams','msfamily','WebViewHost','Clipchamp' |foreach {kill -name $_ -force -ea 0}
-
-## clear win32 uninstall block
-foreach ($name in $remove_win32) { foreach ($sw in $ALLHIVES) {
-  $key = "$sw\Microsoft\Windows\CurrentVersion\Uninstall\$name"; if (-not (test-path $key)) {continue}
-  foreach ($val in 'NoRemove','NoModify','NoRepair') {rp $key $val -force -ea 0}
-  foreach ($val in 'ForceRemove','Delete') {sp $key $val 1 -type Dword -force}
-}}
-PREPARE_EDGE
-
-## find all Edge setup.exe and gather BHO paths for OpenWebSearch / MSEdgeRedirect usage
-$edges = @(); $bho = @(); $edgeupdates = @(); 'LocalApplicationData','ProgramFilesX86','ProgramFiles' |foreach {
-  $folder = [Environment]::GetFolderPath($_); $bho += dir "$folder\Microsoft\Edge*\ie_to_edge_stub.exe" -rec -ea 0
-  if ($WEBV) {$edges += dir "$folder\Microsoft\Edge*\setup.exe" -rec -ea 0 |where {$_ -like '*EdgeWebView*'}}
-  $edges += dir "$folder\Microsoft\Edge*\setup.exe" -rec -ea 0 |where {$_ -notlike '*EdgeWebView*'}
-  $edgeupdates += dir "$folder\Microsoft\EdgeUpdate\*.*.*.*\MicrosoftEdgeUpdate.exe" -rec -ea 0
+if ($InstallEdge -and $UninstallEdge) {
+	throw "You can't use both -InstallEdge and -UninstallEdge as arguments."
 }
 
-## export OpenWebSearch innovative redirector - used by MSEdgeRedirect as well
-$DIR = "$env:SystemDrive\Scripts"; mkdir $DIR -ea 0 >''
-foreach ($b in $bho) { if (test-path $b) { try {copy $b "$DIR\ie_to_edge_stub.exe" -force -ea 0} catch{} } }
-## -------------------------------------------------------------------------------------------------------------------------------
-
-## 4 remove found *Edge* appx packages with unblock tricks
-$provisioned = get-appxprovisionedpackage -online; $appxpackage = get-appxpackage -allusers; $eol = @()
-$store = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Appx\AppxAllUserStore'
-$users = @('S-1-5-18'); if (test-path $store) {$users += $((dir $store -ea 0 |where {$_ -like '*S-1-5-21*'}).PSChildName)}
-foreach ($choice in $remove_appx) { if ('' -eq $choice.Trim()) {continue}
-  foreach ($appx in $($provisioned |where {$_.PackageName -like "*$choice*"})) {
-    $next = !1; foreach ($no in $skip) {if ($appx.PackageName -like "*$no*") {$next = !0}} ; if ($next) {continue}
-    $PackageName = $appx.PackageName; $PackageFamilyName = ($appxpackage |where {$_.Name -eq $appx.DisplayName}).PackageFamilyName 
-    ni "$store\Deprovisioned\$PackageFamilyName" -force >''; $PackageFamilyName  
-    foreach ($sid in $users) {ni "$store\EndOfLife\$sid\$PackageName" -force >''} ; $eol += $PackageName
-    dism /online /set-nonremovableapppolicy /packagefamily:$PackageFamilyName /nonremovable:0 >''
-    remove-appxprovisionedpackage -packagename $PackageName -online -allusers >''
-  }
-  foreach ($appx in $($appxpackage |where {$_.PackageFullName -like "*$choice*"})) {
-    $next = !1; foreach ($no in $skip) {if ($appx.PackageFullName -like "*$no*") {$next = !0}} ; if ($next) {continue}
-    $PackageFullName = $appx.PackageFullName; 
-    ni "$store\Deprovisioned\$appx.PackageFamilyName" -force >''; $PackageFullName
-    foreach ($sid in $users) {ni "$store\EndOfLife\$sid\$PackageFullName" -force >''} ; $eol += $PackageFullName
-    dism /online /set-nonremovableapppolicy /packagefamily:$PackageFamilyName /nonremovable:0 >''
-    remove-appxpackage -package $PackageFullName -allusers >''
-  }
+function Pause ($message = "Press Enter to exit") {
+	if (!$NonInteractive) { $null = Read-Host $message }
 }
-## -------------------------------------------------------------------------------------------------------------------------------
 
-## 5 run found *Edge* setup.exe with uninstall args and wait in-between
-foreach ($setup in $edges) { if (-not (test-path $setup)) {continue}
-  if ($setup -like '*EdgeWebView*') {$target = "--msedgewebview"} else {$target = "--msedge"}
-  $sulevel = ('--system-level','--user-level')[$setup -like '*\AppData\Local\*']
-  $removal = "--uninstall $target $sulevel --verbose-logging --force-uninstall"
-  try {write-host $setup $removal; start -wait $setup -args $removal} catch {}
-  do {sleep 3} while ((get-process -name 'setup','MicrosoftEdge*' -ea 0).Path -like '*\Microsoft\Edge*')
+enum LogLevel {
+	Success
+	Info
+	Warning
+	Error
+	Critical
 }
-## -------------------------------------------------------------------------------------------------------------------------------
+function Write-Status {
+	param (
+		[Parameter(Mandatory=$true)]
+		[ValidateNotNullOrEmpty()]
+		[string]$Text,
+		[LogLevel]$Level = "Info",
+		[switch]$Exit,
+		[string]$ExitString = "Press Enter to exit",
+		[int]$ExitCode = 1 
+	)
 
-## msi installers cleanup
-gp 'HKLM:\SOFTWARE\Classes\Installer\Products\*' 'ProductName' |where {$_.ProductName -like '*Microsoft Edge*'} |foreach { 
-  $prod = ($_.PSChildName -split '(.{8})(.{4})(.{4})(.{4})' -join '-').trim('-')
-  $sort = 7,6,5,4,3,2,1,0,8,12,11,10,9,13,17,16,15,14,18,20,19,22,21,23,25,24,27,26,29,28,31,30,33,32,35,34
-  $code = '{' + -join ($sort |foreach {$prod[$_]}) + '}'; start -wait msiexec.exe -args "/X$code /qn" 2>''
-  ri $_.PSPath -recurse -force
-  foreach ($sw in $ALLHIVES) {ri "$sw\Microsoft\Windows\CurrentVersion\Uninstall\$code" -recurse -force}  
-} 
+	$colour = @(
+		"Green",
+		"White",
+		"Yellow",
+		"Red",
+		"Red"
+	)[$([LogLevel].GetEnumValues().IndexOf($Level))]
 
-## 6 edgeupdate graceful cleanup
-if ($WEBV) {
-  foreach ($sw in $ALLHIVES) {ri "$sw\Microsoft\EdgeUpdate" -recurse -force}  
-  foreach ($UPDT in $edgeupdates) { 
-    if (test-path $UPDT) {write-host "$UPDT /unregsvc";  start -wait $UPDT -args '/unregsvc'}
-    do {sleep 3} while ((get-process -name 'setup','MicrosoftEdge*' -ea 0).Path -like '*\Microsoft\Edge*')
-    if (test-path $UPDT) {write-host "$UPDT /uninstall"; start -wait $UPDT -args '/uninstall'}
-    do {sleep 3} while ((get-process -name 'setup','MicrosoftEdge*' -ea 0).Path -like '*\Microsoft\Edge*')
-  }
-  Unregister-ScheduledTask -TaskName MicrosoftEdgeUpdate* -Confirm:$false -ea 0; ri "$PROGRAMS\Microsoft\Temp" -recurse -force
-} 
-$appdata = $([Environment]::GetFolderPath('ApplicationData'))
-ri "$appdata\Microsoft\Internet Explorer\Quick Launch\User Pinned\TaskBar\Tombstones\Microsoft Edge.lnk" -force
-ri "$appdata\Microsoft\Internet Explorer\Quick Launch\Microsoft Edge.lnk" -force
+	$Text -split "`n" | ForEach-Object {
+		Write-Host "[$($Level.ToString().ToUpper())] $_" -ForegroundColor $colour
+	}
 
-## undo eol unblock trick to prevent latest cumulative update (LCU) failing 
-foreach ($sid in $users) { foreach ($PackageName in $eol) {ri "$store\EndOfLife\$sid\$PackageName" -force >''} }
-
-## .i. "Update policies are configured but will be ignored because this device isn't domain joined" .i.
-$uids = @($EDGE_UID); $cdps = @('msedge'); if ($WEBV) {$uids += $WEBV_UID; $cdps += 'msedgewebview'} 
-foreach ($sw in $ALLHIVES) {
-  sp "$sw\Microsoft\EdgeUpdate" 'DoNotUpdateToEdgeWithChromium' 1 -type Dword -force
-  sp "$sw\Microsoft\EdgeUpdate" 'UpdaterExperimentationAndConfigurationServiceControl' 0 -type Dword -force
-  sp "$sw\Microsoft\EdgeUpdate" 'InstallDefault' 0 -type Dword -force
-  foreach ($uid in $uids) {  
-    sp "$sw\Microsoft\EdgeUpdate" "Install${uid}" 0 -type Dword -force
-    sp "$sw\Microsoft\EdgeUpdate" "EdgePreview${uid}" 0 -type Dword -force
-    sp "$sw\Microsoft\EdgeUpdate" "Update${uid}" 2 -type Dword -force
-    foreach ($trigger in 'on-os-upgrade','on-logon','on-logon-autolaunch','on-logon-startup-boost') {
-      sp "$sw\Microsoft\EdgeUpdate\Clients\${uid}\Commands\$trigger" 'AutoRunOnLogon' 0 -type Dword -force
-      sp "$sw\Microsoft\EdgeUpdate\Clients\${uid}\Commands\$trigger" 'AutoRunOnOSUpgrade' 0 -type Dword -force
-      sp "$sw\Microsoft\EdgeUpdate\Clients\${uid}\Commands\$trigger" 'Enabled' 0 -type Dword -force
-    }
-  }
-  sp "$sw\Microsoft\MicrosoftEdge\Main" 'AllowPrelaunch' 0 -type Dword -force
-  sp "$sw\Microsoft\MicrosoftEdge\TabPreloader" 'AllowTabPreloading' 0 -type Dword -force
-  ## microsoft has no shame, so we are gonna insist opting-out of unsolicited reinstalls with windows updates
-  foreach ($cdp in $cdps) { foreach ($arch in 'x64','x86') { foreach ($zdp in '','-zdp') {
-    sp "$sw\Microsoft\EdgeUpdateDev\CdpNames" "$cdp-stable-win-$arch$zdp" "$cdp-stable-win-arm64$zdp" -force
-  }}}
+	if ($Exit) {
+		Write-Output ""
+		Pause $ExitString
+		exit $ExitCode
+	}
 }
-## -------------------------------------------------------------------------------------------------------------------------------
 
-## 7 add bundled OpenWebSearch script to redirect microsoft-edge: anti-competitive links to the default browser
-$MSEP = ($env:ProgramFiles,${env:ProgramFiles(x86)})[[Environment]::Is64BitOperatingSystem] + '\Microsoft\Edge\Application'
-$IFEO = 'HKLM:\SOFTWARE\Microsoft\Windows NT\CurrentVersion\Image File Execution Options'
-$MIN = ('--headless','--width 1 --height 1')[([environment]::OSVersion.Version.Build) -gt 25179]
-$CMD = "$env:systemroot\system32\conhost.exe $MIN" # AveYo: minimize prompt - see Terminal issue #13914
-ni "HKLM:\SOFTWARE\Classes\microsoft-edge\shell\open\command" -force >''
-sp "HKLM:\SOFTWARE\Classes\microsoft-edge" '(Default)' 'URL:microsoft-edge' -force
-sp "HKLM:\SOFTWARE\Classes\microsoft-edge" 'URL Protocol' '' -force
-sp "HKLM:\SOFTWARE\Classes\microsoft-edge" 'NoOpenWith' '' -force
-sp "HKLM:\SOFTWARE\Classes\microsoft-edge\shell\open\command" '(Default)' "`"$DIR\ie_to_edge_stub.exe`" %1" -force
-ni "HKLM:\SOFTWARE\Classes\MSEdgeHTM\shell\open\command" -force >''
-sp "HKLM:\SOFTWARE\Classes\MSEdgeHTM" 'NoOpenWith' '' -force
-sp "HKLM:\SOFTWARE\Classes\MSEdgeHTM\shell\open\command" '(Default)' "`"$DIR\ie_to_edge_stub.exe`" %1" -force
-ni "$IFEO\ie_to_edge_stub.exe\0" -force >''
-sp "$IFEO\ie_to_edge_stub.exe" 'UseFilter' 1 -type Dword -force
-sp "$IFEO\ie_to_edge_stub.exe\0" 'FilterFullPath' "$DIR\ie_to_edge_stub.exe" -force
-sp "$IFEO\ie_to_edge_stub.exe\0" 'Debugger' "$CMD $DIR\OpenWebSearch.cmd" -force
-ni "$IFEO\msedge.exe\0" -force >''
-sp "$IFEO\msedge.exe" 'UseFilter' 1 -type Dword -force
-sp "$IFEO\msedge.exe\0" 'FilterFullPath' "$MSEP\msedge.exe" -force
-sp "$IFEO\msedge.exe\0" 'Debugger' "$CMD $DIR\OpenWebSearch.cmd" -force
-## new: automatically re-create the needed hardlink if edge is reinstalled
-$ta = New-ScheduledTaskAction -Execute '%Temp%\OpenWebSearchRepair.cmd'
-$tt = New-ScheduledTaskTrigger -Once -At 00:00; $ts = New-ScheduledTaskSettingsSet -AllowStartIfOnBatteries
-Register-ScheduledTask -TaskName 'OpenWebSearchRepair' -Action $ta -Trigger $tt -Settings $ts -RunLevel Highest -Force >''
+function InternetCheck {
+	if ($(Test-Connection "microsoft.com" -Count 1 -EA 0; $?) -eq $false) {
+		Write-Status "Failed to ping Microsoft! You must have an internet connection to reinstall Edge and its components." -Level Critical -Exit -ExitCode 404
+	}
+}
 
-$OpenWebSearch = @'
-@title OpenWebSearch Redux & echo off & set ?= open start menu web search, widgets links or help in your chosen browser - by AveYo
-for /f %%E in ('"prompt $E$S& for %%e in (1) do rem"') do echo;%%E[2t 2>nul & rem AveYo: minimize prompt
-call :reg_var "HKCU\SOFTWARE\Microsoft\Windows\Shell\Associations\UrlAssociations\https\UserChoice" ProgID ProgID
-if /i "%ProgID%" equ "MSEdgeHTM" echo;Default browser is set to Edge! Change it or remove OpenWebSearch script. & pause & exit /b
-call :reg_var "HKCR\%ProgID%\shell\open\command" "" Browser
-set Choice=& for %%. in (%Browser%) do if not defined Choice set "Choice=%%~."
-call :reg_var "HKCR\MSEdgeMHT\shell\open\command" "" FallBack
-set "Edge=" & for %%. in (%FallBack%) do if not defined Edge set "Edge=%%~."
-set "URI=" & set "URL=" & set "NOOP=" & set "PassTrough=%Edge:msedge=edge%"
-set "CLI=%CMDCMDLINE:"=``% "
-if defined CLI set "CLI=%CLI:*ie_to_edge_stub.exe`` =%"
-if defined CLI set "CLI=%CLI:*ie_to_edge_stub.exe =%"
-if defined CLI set "CLI=%CLI:*msedge.exe`` =%"
-if defined CLI set "CLI=%CLI:*msedge.exe =%"
-set "FIX=%CLI:~-1%"
-if defined CLI if "%FIX%"==" " set "CLI=%CLI:~0,-1%"
-if defined CLI set "RED=%CLI:microsoft-edge=%"
-if defined CLI set "URL=%CLI:http=%"
-if defined CLI set "ARG=%CLI:``="%"
-if "%CLI%" equ "%RED%" (set NOOP=1) else if "%CLI%" equ "%URL%" (set NOOP=1)
-if defined NOOP if not exist "%PassTrough%" echo;@mklink /h "%PassTrough%" "%Edge%" >"%Temp%\OpenWebSearchRepair.cmd"
-if defined NOOP if not exist "%PassTrough%" schtasks /run /tn OpenWebSearchRepair 2>nul >nul
-if defined NOOP if not exist "%PassTrough%" timeout /t 3 >nul
-if defined NOOP if exist "%PassTrough%" start "" "%PassTrough%" %ARG%
-if defined NOOP exit /b
-set "URL=%CLI:*microsoft-edge=%"
-set "URL=http%URL:*http=%"
-set "FIX=%URL:~-2%"
-if defined URL if "%FIX%"=="``" set "URL=%URL:~0,-2%"
-call :dec_url
-start "" "%Choice%" "%URL%"
-exit
+function DeleteIfExist($Path) {
+	if (Test-Path $Path) {
+		Remove-Item -Path $Path -Force -Recurse -Confirm:$false
+	}
+}
 
-:reg_var [USAGE] call :reg_var "HKCU\Volatile Environment" value-or-"" variable [extra options]
-set {var}=& set {reg}=reg query "%~1" /v %2 /z /se "," /f /e& if %2=="" set {reg}=reg query "%~1" /ve /z /se "," /f /e
-for /f "skip=2 tokens=* delims=" %%V in ('%{reg}% %4 %5 %6 %7 %8 %9 2^>nul') do if not defined {var} set "{var}=%%V"
-if not defined {var} (set {reg}=& set "%~3="& exit /b) else if %2=="" set "{var}=%{var}:*)    =%"& rem AveYo: v3
-if not defined {var} (set {reg}=& set "%~3="& exit /b) else set {reg}=& set "%~3=%{var}:*)    =%"& set {var}=& exit /b
+function Get-MsiexecAppByName {
+	param(
+		[Parameter(Mandatory = $true)]
+		[ValidateNotNullOrEmpty()]
+		[string]$Name
+	)
 
-:dec_url brute url percent decoding by AveYo
-set ".=%URL:!=}%"&setlocal enabledelayedexpansion& rem brute url percent decoding
-set ".=!.:%%={!" &set ".=!.:{3A=:!" &set ".=!.:{2F=/!" &set ".=!.:{3F=?!" &set ".=!.:{23=#!" &set ".=!.:{5B=[!" &set ".=!.:{5D=]!"
-set ".=!.:{40=@!"&set ".=!.:{21=}!" &set ".=!.:{24=$!" &set ".=!.:{26=&!" &set ".=!.:{27='!" &set ".=!.:{28=(!" &set ".=!.:{29=)!"
-set ".=!.:{2A=*!"&set ".=!.:{2B=+!" &set ".=!.:{2C=,!" &set ".=!.:{3B=;!" &set ".=!.:{3D==!" &set ".=!.:{25=%%!"&set ".=!.:{20= !"
-set ".=!.:{=%%!" &rem set ",=!.:%%=!" & if "!,!" neq "!.!" endlocal& set "URL=%.:}=!%" & call :dec_url
-endlocal& set "URL=%.:}=!%" & exit /b
-rem done
+	$uninstallKeyPath = "Microsoft\Windows\CurrentVersion\Uninstall"
+	$uninstallKeys = (Get-ChildItem -Path @(
+		"HKLM:\SOFTWARE\$uninstallKeyPath",
+		"HKLM:\SOFTWARE\WOW6432Node\$uninstallKeyPath",
+		"HKCU:\SOFTWARE\$uninstallKeyPath",
+		"HKCU:\SOFTWARE\WOW6432Node\$uninstallKeyPath"
+	) -EA SilentlyContinue) -match "\{\b[A-Fa-f0-9]{8}(?:-[A-Fa-f0-9]{4}){3}-[A-Fa-f0-9]{12}\b\}"
 
-'@
-[io.file]::WriteAllText("$DIR\OpenWebSearch.cmd", $OpenWebSearch)
-## -------------------------------------------------------------------------------------------------------------------------------
+	$edges = @()
+	foreach ($key in $uninstallKeys.PSPath) {
+		if (((Get-ItemProperty -Path $key).DisplayName -like "*$Name*") -and ((Get-ItemProperty -Path $key).UninstallString -like "*MsiExec.exe*")) {
+			$edges += Split-Path -Path $key -Leaf
+		}
+	}
 
-## 8 done
-$done = gp 'Registry::HKEY_Users\S-1-5-21*\Volatile*' Edge_Removal -ea 0; if ($done) {rp $done.PSPath Edge_Removal -force -ea 0}
-## -------------------------------------------------------------------------------------------------------------------------------
+	return $edges
+}
+
+# True if it's installed
+function EdgeInstalled {
+	Test-Path $msedgeExe
+}
+
+function KillEdgeProcesses {
+	$ErrorActionPreference = 'SilentlyContinue'
+	foreach ($service in (Get-Service -Name "*edge*" | Where-Object {$_.DisplayName -like "*Microsoft Edge*"}).Name) {
+		Stop-Service -Name $service -Force
+	}
+	foreach (
+		$process in
+		(Get-Process | Where-Object {($_.Path -like "$([Environment]::GetFolderPath('ProgramFilesX86'))\Microsoft\*") -or ($_.Name -like "*msedge*")}).Id
+	) {
+		Stop-Process -Id $process -Force
+	}
+	$ErrorActionPreference = 'Continue'	
+}
+
+function RemoveEdgeChromium([bool]$AlreadyUninstalled) {
+	Write-Status -Text "Trying to find Edge uninstallers..."
+
+	# get Edge MsiExec uninstallers
+	# commonly installed with WinGet (it installs the Enterprise MSI)
+	$msis = Get-MsiexecAppByName -Name "Microsoft Edge"
+
+	# find using common locations - used as a backup
+	function UninstallStringFail {
+		if ($msis.Count -le 0) {
+			Write-Status -Text "Couldn't parse uninstall string for Edge. Trying to find uninstaller manually." -Level Warning
+		}
+
+		$script:edgeUninstallers = @()
+		'LocalApplicationData','ProgramFilesX86','ProgramFiles' | ForEach-Object {
+			$folder = [Environment]::GetFolderPath($_)
+			$script:edgeUninstallers += Get-ChildItem "$folder\Microsoft\Edge*\setup.exe" -Recurse -EA 0 |
+				Where-Object {($_ -like '*Edge\Application*') -or ($_ -like '*SxS\Application*')}
+		}
+	}
+
+	# find using Registry
+	$uninstallKeyPath = "$baseKey\Windows\CurrentVersion\Uninstall\Microsoft Edge"
+	$uninstallString = (Get-ItemProperty -Path $uninstallKeyPath -EA 0).UninstallString
+	if ([string]::IsNullOrEmpty($uninstallString) -and ($msis.Count -le 0)) {
+		$uninstallString = $null
+		UninstallStringFail
+	} else {
+		# split uninstall string for path & args
+		$uninstallPath, $uninstallArgs = $uninstallString -split '"', 3 |
+			Where-Object { $_ } |
+			ForEach-Object { [System.Environment]::ExpandEnvironmentVariables($_.Trim()) }
+
+		# check if fully qualified (should normally be), otherwise it could be null or something in the working dir
+		if (![System.IO.Path]::IsPathRooted($uninstallPath) -or !(Test-Path $uninstallPath -PathType Leaf)) {
+			$uninstallPath = $null
+			UninstallStringFail
+		}
+	}
+
+	# throw if installers aren't found
+	if (($msis.Count -le 0) -and ($script:edgeUninstallers.Count -le 0) -and !$uninstallPath) {
+		$uninstallError = @{
+			Text = "Failed to find uninstaller! " + $(if ($AlreadyUninstalled) {
+				"This likely means Edge is already uninstalled."
+			} else {
+				"The uninstall can't continue. :("
+			})
+			Level = if ($AlreadyUninstalled) { 'Warning' } else { 'Critical' }
+			Exit = $true
+			ExitCode = 2
+		}
+		Write-Status @uninstallError
+	} else {
+		Write-Status "Found Edge uninstallers."
+	}
+
+	# toggles an EU region - this is because anyone in the EEA can uninstall Edge
+	# this key is checked by the Edge uninstaller
+	function ToggleEURegion([bool]$Enable) {
+		$geoKey = "Registry::HKEY_USERS\.DEFAULT\Control Panel\International\Geo"
+
+		# sets Geo to France, which is in the EEA
+		$values = @{
+			"Name" = "FR"
+			"Nation" = "84"
+		}
+		$geoChange = 'EdgeSaved'
+
+		if ($Enable) {
+			$values.GetEnumerator() | ForEach-Object {
+				Rename-ItemProperty -Path $geoKey -Name $_.Key -NewName "$($_.Key)$geoChange" -Force
+				Set-ItemProperty -Path $geoKey -Name $_.Key -Value $_.Value -Force
+			}
+		} else {
+			$values.GetEnumerator() | ForEach-Object {
+				Remove-ItemProperty -Path $geoKey -Name $_.Key -Force -EA 0
+				Rename-ItemProperty -Path $geoKey -Name "$($_.Key)$geoChange" -NewName $_.Key -Force -EA 0
+			}
+		}
+	}
+
+	function ModifyRegionJSON {
+		$cleanup = $false
+		$script:integratedServicesPath = "$sys32\IntegratedServicesRegionPolicySet.json"
+
+		if (Test-Path $integratedServicesPath) {
+			$cleanup = $true
+			try {
+				$admin = [System.Security.Principal.NTAccount]$(New-Object System.Security.Principal.SecurityIdentifier('S-1-5-32-544')).Translate([System.Security.Principal.NTAccount]).Value
+
+				# get perms (normally TrustedInstaller only)
+				$acl = Get-Acl -Path $integratedServicesPath
+				$script:backup = [System.Security.AccessControl.FileSecurity]::new()
+				$script:backup.SetSecurityDescriptorSddlForm($acl.Sddl)
+				# full control
+				$acl.SetOwner($admin)
+				$rule = New-Object System.Security.AccessControl.FileSystemAccessRule($admin, "FullControl", "Allow")
+				$acl.AddAccessRule($rule)
+				# set modified ACL
+				Set-Acl -Path $integratedServicesPath -AclObject $acl
+
+				# modify the stuff
+				$integratedServices = Get-Content $integratedServicesPath | ConvertFrom-Json
+				($integratedServices.policies | Where-Object { ($_.'$comment' -like "*Edge*") -and ($_.'$comment' -like "*uninstall*") }).defaultState = 'enabled'
+				$modifiedJson = $integratedServices | ConvertTo-Json -Depth 100
+
+				$script:backupIntegratedServicesName = "IntegratedServicesRegionPolicySet.json.$([System.IO.Path]::GetRandomFileName())"
+				Rename-Item $integratedServicesPath -NewName $script:backupIntegratedServicesName -Force
+				Set-Content $integratedServicesPath -Value $modifiedJson -Force -Encoding UTF8
+			} catch {
+				Write-Error "Failed to modify region policies. $_"
+			}
+		} else {
+			Write-Status -Text "'$integratedServicesPath' not found." -Level Warning
+		}
+
+		return $cleanup
+	}
+
+
+	# Edge uninstalling logic
+	function UninstallEdge {
+		# MSI packages have to be uninstalled first, otherwise it breaks
+		foreach ($msi in $msis) {
+			Write-Status 'Uninstalling Edge using Windows Installer...'
+			Start-Process -FilePath "msiexec.exe" -ArgumentList "/qn /X$(Split-Path -Path $msi -Leaf) REBOOT=ReallySuppress /norestart" -Wait
+		}
+
+		# uninstall standard Edge installs
+		if ($uninstallPath) {  # found from Registry
+			Start-Process -Wait -FilePath $uninstallPath -ArgumentList "$uninstallArgs --force-uninstall" -WindowStyle Hidden
+		} else {  # found from system files
+			foreach ($setup in $edgeUninstallers) {
+				if (Test-Path $setup) {
+					$sulevel = ('--system-level','--user-level')[$setup -like '*\AppData\Local\*']
+					Start-Process -Wait $setup -ArgumentList "--uninstall --msedge $sulevel --channel=stable --verbose-logging --force-uninstall"
+				}
+			}
+		}
+
+		# return if Edge is installed or not
+		return EdgeInstalled
+	}
+
+	# things that should always be done before uninstall
+	function GlobalRemoveMethods {
+		Write-Status "Using method $method..." -Level Warning
+
+		# delete experiment_control_labels for key that prevents (or prevented) uninstall
+		Remove-ItemProperty -Path "$baseKey\EdgeUpdate\ClientState\{56EB18F8-B008-4CBD-B6D2-8C97FE7E9062}" -Name 'experiment_control_labels' -Force -EA 0
+
+		# allow Edge uninstall
+		$devKeyPath = "$baseKey\EdgeUpdateDev"
+		if (!(Test-Path $devKeyPath)) { New-Item -Path $devKeyPath -ItemType "Key" -Force | Out-Null }
+		Set-ItemProperty -Path $devKeyPath -Name "AllowUninstall" -Value "" -Type String -Force
+	
+		Write-Status "Terminating Microsoft Edge processes..."
+		KillEdgeProcesses
+	}
+
+	# go through each uninstall method
+	# yes, i'm aware this seems excessive, but i'm just trying to make sure it works on the most installs possible
+	# it does bloat the script lots though... i'll clean it up in a future release, but for now, i'm just fixing it
+	$fail = $true
+	$method = 1
+	function CleanupMsg { Write-Status "Cleaning up after method $method..." }
+	while ($fail) {
+		switch ($method) {
+			# makes Edge think the old legacy UWP is still installed
+			# seems to fail on some installs?
+			1 {
+				GlobalRemoveMethods
+				if (!(Test-Path "$edgeUWP\MicrosoftEdge.exe")) {
+					New-Item $edgeUWP -ItemType Directory -ErrorVariable cleanup -EA 0 | Out-Null
+					New-Item "$edgeUWP\MicrosoftEdge.exe" -EA 0 | Out-Null
+					$cleanup = $true
+				}
+
+				# attempt uninstall
+				$fail = UninstallEdge
+
+				if ($cleanup) {
+					CleanupMsg
+					Remove-Item $edgeUWP -Force -EA 0 -Recurse
+				}
+			}
+
+			# not having windir defined is a condition to allow uninstall
+			# found in the strings of the setup ^
+			2 {
+				GlobalRemoveMethods
+				$envPath = 'HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager\Environment'
+				try {
+					# delete windir variable temporarily
+					Set-ItemProperty -Path $envPath -Name 'windir' -Value '' -Type ExpandString
+					$env:windir = [System.Environment]::GetEnvironmentVariable('windir', [System.EnvironmentVariableTarget]::Machine)
+
+					# attempt uninstall
+					$fail = UninstallEdge
+				} finally {
+					CleanupMsg
+					# this is the default
+					Set-ItemProperty -Path $envPath -Name 'windir' -Value '%SystemRoot%' -Type ExpandString
+				}
+			}
+
+			# changes region in Registry
+			# currently not known to work, kept for legacy reasons
+			3 {
+				GlobalRemoveMethods
+				ToggleEURegion $true
+
+				$fail = UninstallEdge
+
+				CleanupMsg
+				ToggleEURegion $false
+			}
+
+			# modifies IntegratedServicesRegionPolicySet to add current region to allow list
+			# currently not known to work, kept for legacy reasons
+			4 {
+				GlobalRemoveMethods
+				$cleanup = ModifyRegionJSON
+				
+				# attempt uninstall
+				$fail = UninstallEdge
+
+				# cleanup
+				if ($cleanup) {
+					CleanupMsg
+					Remove-Item $integratedServicesPath -Force
+					Rename-Item "$sys32\$backupIntegratedServicesName" -NewName $integratedServicesPath -Force
+					Set-Acl -Path $integratedServicesPath -AclObject $backup
+				}
+			}
+
+			# everything fails ╰（‵□′）╯
+			default {
+				Write-Status "The uninstall methods failed for the Edge installers found. Nothing else can be done." -Level Critical -Exit -ExitCode 3
+			}
+		}
+
+		$method++
+	}
+	Write-Status "Successfully removed Edge! :)" -Level Success
+
+	# remove old shortcuts
+	"$([Environment]::GetFolderPath('Desktop'))\Microsoft Edge.lnk",
+	"$([Environment]::GetFolderPath('CommonStartMenu'))\Microsoft Edge.lnk" | ForEach-Object { DeleteIfExist $_ }
+
+	# restart explorer if Copilot is enabled - this will hide the Copilot button
+	if ((Get-ItemProperty -Path "HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\Advanced" -Name "ShowCopilotButton" -EA 0)."ShowCopilotButton" -eq 1) {
+		Stop-Process -Name explorer -Force -ErrorAction SilentlyContinue
+	}
+}
+
+function RemoveEdgeAppX {
+	# i'm aware of how this is deprecated
+	# kept for legacy purposes just in case someone's using an older build of Windows
+
+	$SID = (New-Object System.Security.Principal.NTAccount([Environment]::UserName)).Translate([Security.Principal.SecurityIdentifier]).Value
+
+	# remove from Registry
+	$appxStore = '\SOFTWARE\Microsoft\Windows\CurrentVersion\Appx\AppxAllUserStore'
+	$pattern = "HKLM:$appxStore\InboxApplications\Microsoft.MicrosoftEdge_*_neutral__8wekyb3d8bbwe"
+	$edgeAppXKey = (Get-Item -Path $pattern).PSChildName
+	if (Test-Path "$pattern") { reg delete "HKLM$appxStore\InboxApplications\$edgeAppXKey" /f | Out-Null }
+
+	# make the Edge AppX able to uninstall and uninstall
+	New-Item -Path "HKLM:$appxStore\EndOfLife\$SID\Microsoft.MicrosoftEdge_8wekyb3d8bbwe" -Force | Out-Null
+	Get-AppxPackage -Name Microsoft.MicrosoftEdge | Remove-AppxPackage | Out-Null
+	Remove-Item -Path "HKLM:$appxStore\EndOfLife\$SID\Microsoft.MicrosoftEdge_8wekyb3d8bbwe" -Force | Out-Null
+}
+
+
+function InstallEdgeChromium {
+	InternetCheck
+
+	$temp = mkdir (Join-Path $([System.IO.Path]::GetTempPath()) $(New-Guid))
+	$msi = "$temp\edge.msi"
+	$msiLog = "$temp\edgeMsi.log"
+	
+	if ([Environment]::Is64BitOperatingSystem) {
+		$arm = ((Get-CimInstance -Class Win32_ComputerSystem).SystemType -match 'ARM64') -or ($env:PROCESSOR_ARCHITECTURE -eq 'ARM64')
+		$archString = ('x64', 'arm64')[$arm]
+	} else {
+		$archString = 'x86'
+	}
+	
+	Write-Status "Requesting from the Microsoft Edge Update API..."
+	try {
+		try {
+			$edgeUpdateApi = (Invoke-WebRequest "https://edgeupdates.microsoft.com/api/products" -UseBasicParsing).Content | ConvertFrom-Json
+		} catch {
+			Write-Status "Failed to request from EdgeUpdate API!
+Error: $_" -Level Critical -Exit -ExitCode 4
+		}
+
+		$edgeItem = ($edgeUpdateApi | ? { $_.Product -eq 'Stable' }).Releases |
+			Where-Object { $_.Platform -eq 'Windows' -and $_.Architecture -eq $archString } |
+			Where-Object { $_.Artifacts.Count -ne 0 } | Select-Object -First 1
+		
+		if ($null -eq $edgeItem) {
+			Write-Status "Failed to parse EdgeUpdate API! No matching artifacts found." -Level Critical -Exit
+		}
+
+		$hashAlg = $edgeItem.Artifacts.HashAlgorithm | % { if ([string]::IsNullOrEmpty($_)) { "SHA256" } else { $_ } }
+		foreach ($var in @{
+			link = $edgeItem.Artifacts.Location
+			hash = $edgeItem.Artifacts.Hash
+			version = $edgeItem.ProductVersion
+			sizeInMb = [math]::round($edgeItem.Artifacts.SizeInBytes/1Mb)
+			released = Get-Date $edgeItem.PublishedTime
+		}.GetEnumerator()) {
+			$val = $var.Value | Select-Object -First 1
+			if ($val.Length -le 0) {
+				Set-Variable -Name $var.Key -Value "Undefined"
+				if ($var.Key -eq 'link') { throw "Failed to parse download link!" }
+			} else {
+				Set-Variable -Name $var.Key -Value $val
+			}
+		}
+	} catch {
+		Write-Status "Failed to parse Microsoft Edge from `"$link`"!
+Error: $_" -Level Critical -Exit -ExitCode 5
+	}
+	Write-Status "Parsed Microsoft Edge Update API!" -Level Success
+	
+	Write-Host "`nDownloading Microsoft Edge:" -ForegroundColor Cyan
+	@(
+		@("Released on: ", $released),
+		@("Version: ", "$version (Stable)"),
+		@("Size: ", "$sizeInMb Mb")
+	) | Foreach-Object {
+		Write-Host " - " -NoNewline -ForegroundColor Magenta
+		Write-Host $_[0] -NoNewline -ForegroundColor Yellow
+		Write-Host $_[1]
+	}
+	
+	Write-Output ""
+	try {
+		if ($null -eq (Get-Command curl.exe -EA 0)) {
+			Write-Status "Couldn't find cURL, using Invoke-WebRequest, which is slower..." -Level Warning
+			Invoke-WebRequest -Uri $link -Output $msi -UseBasicParsing
+		} else {
+			curl.exe -#L "$link" -o "$msi"
+		}
+	} catch {
+		Write-Status "Failed to download Microsoft Edge from `"$link`"!
+Error: $_" -Level Critical -Exit -ExitCode 6
+	}
+	Write-Output ""
+	
+	if ($hash -eq "Undefined") {
+		Write-Status "Not verifying hash as it's undefined, download might have failed." -Level Warning
+	} else {
+		Write-Status "Verifying download by checking its hash..."
+		if ((Get-FileHash -LiteralPath $msi -Algorithm $hashAlg).Hash -eq $hash) {
+			Write-Status "Verified the Microsoft Edge installer!" -Level Success
+		} else {
+			Write-Status "Edge installer hash does not match. The installer might be corrupted. Continuing anyways..." -Level Error
+		}
+	}
+	
+	Write-Status "Installing Microsoft Edge..."
+	Start-Process -FilePath "msiexec.exe" -ArgumentList "/i `"$msi`" /l `"$msiLog`" /quiet" -Wait
+	
+	if (!(Test-Path $msiLog)) {
+		Write-Status "Couldn't find installer log at `"$msiLog`"! This likely means it failed." -Level Critical -Exit -ExitCode 7
+	}
+	
+	Write-Status -Text "Installer log path: `"$msiLog`""
+	if ($null -eq ($(Get-Content $msiLog) -like "*Product: Microsoft Edge -- * completed successfully.*")) {
+		Write-Status "Can't find success string from Edge install log - it seems like the install was a failure." -Level Error -Exit -ExitCode 8
+	}
+
+	Write-Status -Text "Installed Microsoft Edge!" -Level Success
+}
+
+function InstallWebView {
+	InternetCheck
+
+	$dlPath = "$((Join-Path $([System.IO.Path]::GetTempPath()) $(New-Guid)))-webview2.exe"
+	$link = "https://go.microsoft.com/fwlink/p/?LinkId=2124703"
+
+	Write-Status "Downloading Edge WebView..."
+	try {
+		if ($null -eq (Get-Command curl.exe -EA 0)) {
+			Write-Status "Couldn't find cURL, using Invoke-WebRequest, which is slower..." -Level Warning
+			Invoke-WebRequest -Uri $link -Output $dlPath -UseBasicParsing
+		} else {
+			curl.exe -Ls "$link" -o "$dlPath"
+		}
+	} catch {
+		Write-Status "Failed to download Edge WebView from `"$link`"!
+Error: $_" -Level Critical -Exit -ExitCode 9
+	}
+
+	Write-Status "Installing Edge WebView..."
+	Start-Process -FilePath "$dlPath" -ArgumentList "/silent /install" -Wait
+
+	Write-Status "Installed Edge WebView!" -Level Success
+}
+
+# SYSTEM check - using SYSTEM previously caused issues
+if ([Security.Principal.WindowsIdentity]::GetCurrent().User.Value -eq 'S-1-5-18') {
+	Write-Status "This script can't be ran as TrustedInstaller/SYSTEM.
+Please relaunch this script under a regular admin account." -Level Critical -Exit
+} else {
+	if (!([Security.Principal.WindowsPrincipal] [Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltinRole]::Administrator)) {
+		if ($PSBoundParameters.Count -le 0 -and !$args) {
+				Start-Process cmd "/c PowerShell -NoP -EP Bypass -File `"$PSCommandPath`"" -Verb RunAs
+				exit
+		} else {
+			throw "This script must be run as an administrator."
+		}
+	}
+}
+
+# main menu
+$edgeInstalled = EdgeInstalled
+if (!$UninstallEdge -and !$InstallEdge -and !$InstallWebView) {
+	$host.UI.RawUI.WindowTitle = "EdgeRemover $version | made by @he3als"	
+
+	$RemoveEdgeData = $false
+	while (!$continue) {
+		Clear-Host
+		$description = "This script removes or installs Microsoft Edge."
+		Write-Host "$description`n" -ForegroundColor Blue
+		Write-Host @"
+To select an option, type its number.
+To perform an action, also type its number.
+"@ -ForegroundColor Yellow
+
+		Write-Host "`nEdge is currently detected as: " -NoNewline -ForegroundColor Green
+		Write-Host "$(@("Uninstalled", "Installed")[$edgeInstalled])" -ForegroundColor Cyan
+
+		Write-Host "`n$("-" * $description.Length)" -ForegroundColor Magenta
+
+		if ($RemoveEdgeData) {$colourData = "Green"; $textData = "Selected"} else {$colourData = "Red"; $textData = "Unselected"}
+
+		Write-Host "`nOptions:"
+		Write-Host "[1] Remove Edge User Data ($textData)" -ForegroundColor $colourData
+		
+		Write-Host "`nActions:"
+		Write-Host @"
+[2] Uninstall Edge
+[3] Install Edge
+[4] Install WebView
+[5] Install both Edge & WebView
+"@ -ForegroundColor Cyan
+
+		$userInput = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown')
+
+		switch ($userInput.VirtualKeyCode) {
+			49 { # remove Edge user data (1)
+				$RemoveEdgeData = !$RemoveEdgeData
+			}
+			50 { # uninstall Edge (2)
+				$UninstallEdge = $true
+				$continue = $true
+			}
+			51 { # reinstall Edge (3)
+				$InstallEdge = $true
+				$continue = $true
+			}
+			52 { # reinstall WebView (4)
+				$InstallWebView = $true
+				$continue = $true
+			}
+			53 { # reinstall both (5)
+				$InstallWebView = $true
+				$InstallEdge = $true
+				$continue = $true
+			}
+		}
+	}
+
+	Clear-Host
+}
+
+if ($UninstallEdge) {
+	Write-Status "Uninstalling Edge Chromium..."
+	RemoveEdgeChromium $(!$edgeInstalled)
+	if ($null -ne (Get-AppxPackage -Name Microsoft.MicrosoftEdge)) {
+		if ($KeepAppX) {
+			Write-Status "AppX Edge is being left, there might be a stub..." -Level Warning
+		} else {
+			Write-Status "Uninstalling AppX Edge..." -Level Warning
+			RemoveEdgeAppx
+		}
+	}
+	Write-Output ""
+}
+
+if ($RemoveEdgeData) {
+	KillEdgeProcesses
+	DeleteIfExist "$([Environment]::GetFolderPath('LocalApplicationData'))\Microsoft\Edge"
+	Write-Status "Removed any existing Edge Chromium user data."
+	Write-Output ""
+}
+
+if ($InstallEdge) {
+	InstallEdgeChromium
+	Write-Output ""
+}
+if ($InstallWebView) {
+	InstallWebView
+	Write-Output ""
+}
+
+Write-Host "Completed." -ForegroundColor Cyan
+if ($NonInteractive) { exit }
+Pause